fix (again) cast

+++ client.go
@@ -221,7 +221,7 @@ func strToFloat(inputString string) (float64, error) {
 func (client *Client) Add(key string, timestamp int64, value float64) (timestamp int64, err error) {
 	conn := client.Pool.Get()
 	defer conn.Close()
-	return redis.Int( conn.Do("TS.ADD", key, timestamp, floatToStr(value)))
+	return redis.Int64( conn.Do("TS.ADD", key, timestamp, floatToStr(value)))
 }
 
 type DataPoint struct {
